+++ src/components/Footer.vue
@@ -1,0 +1,9 @@
+<template>
+  <footer class="footer has-margin-top-30">
+    <div class="columns">
+      <div class="column">
+        <g-link to="/recommendations" class="is-size-7">Recommendations</g-link>
+      </div>
+    </div>
+  </footer>
+</template>