+++ src/main.js
@@ -6,13 +6,12 @@ import Buefy from "buefy";
 import "~/assets/scss/bulma_variables.scss";
 import "bulma-helpers/css/bulma-helpers.min.css";
 
-import Vue from 'vue'
 import axios from 'axios'
 import VueAxios from 'vue-axios'
 
 // Local files
-// Layouts
 import DefaultLayout from "~/layouts/Default.vue";
+import "~/assets/scss/custom.scss";
 
 export default function(Vue, { router, head, isClient }) {
     // Plugins