+++ src/pages/Experience.vue
@@ -1,52 +1,31 @@
 <template>
   <Layout>
-    <div class="title-bar columns is-vcentered is-gapless has-padding-top-20">
-      <div class="column is-2 has-text-centered">
-        <g-image src="~/assets/images/experience.png" width="150px" class="is-round-image" />
-      </div>
-      <div class="column has-margin-left-20">
-        <div class="description has-padding-10 has-padding-left-30 has-padding-bottom-30">
-          <h1
-            class="has-text-weight-bold is-uppercase is-size-6 is-pulled-left has-margin-0"
-          >My Experience</h1>
-          <p
-            class="is-size-7 is-pulled-right is-italic"
-          >My life so far? Still running, started small, going big!</p>
-        </div>
-      </div>
-    </div>
+    <Hero title="My Experience" subtitle="My life so far? Still running, started small, going big!">
+      <g-image src="~/assets/images/experience.png" width="150px" class="is-round-image" />
+    </Hero>
 
-    <div class="experience has-padding-top-30">
-      <div class="container has-margin-top-30">
-        <timeline timeline-bg="#001934">
-          <Experience
-            v-for="experience in $page.experience.edges"
-            :key="experience.id"
-            :experience="experience.node"
-          />
-        </timeline>
-      </div>
+    <div class="container has-margin-top-30">
+      <timeline timeline-bg="#001934">
+        <Experience
+          v-for="experience in $page.experience.edges"
+          :key="experience.id"
+          :experience="experience.node"
+        />
+      </timeline>
     </div>
   </Layout>
 </template>
 
-<style>
-.title-bar .description {
-  background: #091a28;
-  border-radius: 100px 0 0 100px;
-}
-.experience {
-  background: #001934;
-}
-</style>
 
 <script>
 // Components
+import Hero from "~/components/Hero";
 import Experience from "~/components/Experience";
 import { Timeline } from "vue-cute-timeline";
 
 export default {
   components: {
+    Hero,
     Experience,
     Timeline
   },